added iteration to steps sleep and water to show all added inputs

+++ public/js/dashboard.js
@@ -201,25 +201,35 @@ const deleteWaterBtn = document.querySelector('button[form="deleteFormWater"]');
 deleteWaterBtn.addEventListener('click', (event) => {
   event.preventDefault();
   console.log('button clicked:', deleteWaterBtn);
-  const waterLabel = document.querySelector('#waterLabel');
-  const waterId = waterLabel.previousElementSibling.getAttribute('id');
-  console.log(waterId);
 
-  const deleteRequest = axios.delete('/api/dashboard/water/' + waterId);
-  deleteRequest
-    .then(function (response) {
-      console.log('Response:', response.data);
-      // Close the modal
-      let editModal = bootstrap.Modal.getInstance(
-        document.getElementById('modalWater')
-      );
-      editModal.hide();
-      // Reload the page
-      window.location.reload();
-    })
-    .catch(function (error) {
-      console.error('Error:', error.message);
-    });
+  const waterLabels = document.querySelectorAll('.waterLabel');
+
+  for (i = 0; i < waterLabels.length; i ++) {
+    const waterLabel = waterLabels[i];
+    const waterId = waterLabel.previousElementSibling.getAttribute('id');
+    console.log(waterId);
+
+    if (waterLabel.previousElementSibling.checked) {
+      const waterId = waterLabel.previousElementSibling.getAttribute('id');
+      console.log(waterId);
+
+      const deleteRequest = axios.delete('/api/dashboard/water/' + waterId);
+      deleteRequest
+        .then(function (response) {
+          console.log('Response:', response.data);
+          // Close the modal
+          let editModal = bootstrap.Modal.getInstance(
+            document.getElementById('modalWater')
+          );
+          editModal.hide();
+          // Reload the page
+          window.location.reload();
+        })
+        .catch(function (error) {
+          console.error('Error:', error.message);
+        });
+    }
+  }
 });
 
 //* -----------------------------------------------------------------------------
@@ -255,25 +265,32 @@ deleteSleepBtn.addEventListener('click', (event) => {
   event.preventDefault();
   console.log('button clicked:', deleteSleepBtn);
 
-  const sleepLabel = document.querySelector('#sleepLabel');
-  const sleepId = sleepLabel.previousElementSibling.getAttribute('id');
-  console.log(sleepId);
+  const sleepLabels = document.querySelectorAll('.sleepLabel');
 
-  const deleteRequest = axios.delete('/api/dashboard/sleep/' + sleepId);
-  deleteRequest
-    .then(function (response) {
-      console.log('Response:', response.data);
-      // Close the modal
-      let editModal = bootstrap.Modal.getInstance(
-        document.getElementById('modalSleep')
-      );
-      editModal.hide();
-      // Reload the page
-      window.location.reload();
-    })
-    .catch(function (error) {
-      console.error('Error:', error.message);
-    });
+  for (i = 0; i < sleepLabels.length; i ++) {
+    const sleepLabel = sleepLabels[i];
+
+    if (sleepLabel.previousElementSibling.checked) {
+      const sleepId = sleepLabel.previousElementSibling.getAttribute('id');
+      console.log(sleepId);
+
+      const deleteRequest = axios.delete('/api/dashboard/sleep/' + sleepId);
+      deleteRequest
+        .then(function (response) {
+          console.log('Response:', response.data);
+          // Close the modal
+          let editModal = bootstrap.Modal.getInstance(
+            document.getElementById('modalSleep')
+          );
+          editModal.hide();
+          // Reload the page
+          window.location.reload();
+        })
+        .catch(function (error) {
+          console.error('Error:', error.message);
+        });
+      }
+  }
 });
 
 //* -----------------------------------------------------------------------------
@@ -372,21 +389,26 @@ deleteStepsBtn.addEventListener('click', (event) => {
 
   console.log('button clicked:', deleteStepsBtn);
 
-  const stepsLabel = document.querySelector('#stepsLabel');
+  const stepsLabels = document.querySelectorAll('.stepsLabel');
 
-  const stepsId = stepsLabel.previousElementSibling.getAttribute('id');
+  for (i = 0; i < stepsLabels.length; i ++) {
+    const stepsLabel = stepsLabels[i];
 
-  console.log(stepsId);
+    if (stepsLabel.previousElementSibling.checked) {
+    const stepsId = stepsLabel.previousElementSibling.getAttribute('id');
+    console.log(stepsId);
 
-  const deleteRequest = axios.delete('/api/dashboard/steps/' + stepsId);
-  deleteRequest
-    .then(function (response) {
-      console.log('Response:', response.data);
-      window.location.replace('/api/dashboard');
-    })
-    .catch(function (error) {
-      console.error('Error:', error.message);
-    });
+      const deleteRequest = axios.delete('/api/dashboard/steps/' + stepsId);
+      deleteRequest
+        .then(function (response) {
+          console.log('Response:', response.data);
+          window.location.replace('/api/dashboard');
+        })
+        .catch(function (error) {
+          console.error('Error:', error.message);
+        });
+    }
+  }
 });
 
 //* -----------------------------------------------------------------------------
